Variable alpha method
Added a method 3 which allows variable alpha selection.
A maximum and minimum value for alpha needs to be given.
The maximum and minimum magnitudes of diagonal entries of the line Z
components are collected.
For each branch, the value of alpha is linearily interpolated so that
the maximum Z corresponds to the minimum alpha, and the minimum Z to the
maximum alpha.