Implemented variable branch gamma
The matrix used to convert the diagonal bus loads to square matrices,
gamma, can now be branch dependent using `opt.gamma_method  = 1`.

Currently, the lossless distflow will first be solved to get an estimate
of the voltages. The ratio of voltage is then used to weight the
off-diagonal terms in gamma.

NOTE: this does not seem to improve results unfortunately.