Updated single phase distflow function
Function `distflow_lossy` now uses slightely different notation but most
importantly has methods implemented to create per branch `alpha` values
rather than just a scalar.

For the lossless case (all `alpha=0.5`) the two implementations are
completely identical. When `alpha ~= 0.5` however, even when it is a
scalar there seem to be slight numerical differences (order of 1e-16 for
voltage and 1e-14 or 1e-15 for flows (in MW/MVAr) ).

In the event this error does not appear to be strickly numerical, the
old implementation is saved under `distflow_lossy_old`.