Lossy distflow test against matpower feeders
There are a couple of small radial feeders in MATPOWER that are tested
against the parametrization found for the distlow.

Results are somewhat mixed, however, these are very small feeders.

Considering 6 feeders, the average of the rms results is taken and the
max of the inf_norm. The result is that when lossy distflow is worse it
is worse by a little, when its better, it is sometimes better by a lot.

That's the current stand of this file.