Branch output from distflow_multi
Branch powerflows are now calculated and added to the branch structure
array. This is done by first solving for `xi` using the solved voltage
squared vector, constructing `psi` and then selecting the appropriate
diagonal entries.

Note: there may be a better way to calculate the powers but this is a
first attempt.

Also Note: there are some beginnings of a method 11 for approximating
losses implemented these are NOT yet complete.

Finally, alpha_method 2 has been deprecated (anyway there will probably
be a renumbering of methods at some point).