Factored out internal distflow_muti functions
Several functions that were previously internal functions to the
distflow_multi.m file are now externalized as standalone functions.
This is done in an effort to make it possible to export matrices from
the distflow function and then solve multiple scenarios, rather than
have to spend the effor to form the matrices each time.

 Changes to be committed:
	new file:   gamma_phi.m
	new file:   getsigma.m
	new file:   kdiag.m
	new file:   updateKmat.m
	new file:   yload.m